+++ src/main/java/com/cos/blog/dto/ReplySaveRequestDto.java
@@ -7,8 +7,8 @@ import javax.validation.constraints.NotBlank;
 @Getter
 public class ReplySaveRequestDto {
 
-    private int bid;
-    private int uid;
+    private Long bid;
+    private Long uid;
 
     @NotBlank(message = "내용을 입력해주세요.")
     private String content;
